Moved DuckDB reports into their own file.
